backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-generativeai>=0.3.0
nltk>=3.8.0
unidecode>=1.3.0
PyPDF2>=3.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
slowapi>=0.1.9
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.0
